Extraction d'un composant Carte
